+++ R/netclu_labelprop.R
@@ -25,7 +25,7 @@
 #' 
 #' @param return_node_type a `character` indicating what types of nodes
 #' ("sites", "species" or "both") should be returned in the output
-#' (`keep_nodes_type="both"` by default).
+#' (`return_node_type = "both"` by default).
 #' 
 #' @param algorithm_in_output a `boolean` indicating if the original output
 #' of `communities` should be returned in the output (see Value).
@@ -95,8 +95,12 @@ netclu_labelprop <- function(net,
                              return_node_type = "both",
                              algorithm_in_output = TRUE) {
   
-  # Control input net
-  controls(args = NULL, data = net, type = "input_bioregion.pairwise.metric")
+  # Control input net (+ check similarity if not bipartite)
+  controls(args = bipartite, data = NULL, type = "boolean")
+  isbip <- bipartite
+  if(!isbip){
+    controls(args = NULL, data = net, type = "input_similarity")
+  }
   controls(args = NULL, data = net, type = "input_net")
   
   # Control input weight & index
@@ -109,8 +113,6 @@ netclu_labelprop <- function(net,
   }
   
   # Control input bipartite
-  controls(args = bipartite, data = NULL, type = "boolean")
-  isbip <- bipartite
   if (isbip) {
     controls(args = NULL, data = net, type = "input_net_bip")
     controls(args = site_col, data = net, type = "input_net_bip_col")