+++ R/nhclu_clara.R
@@ -1,7 +1,8 @@
-#' Non hierarchical clustering: partitioning around medoids
+#' Non hierarchical clustering: CLARA
 #'
 #' This function performs non hierarchical clustering on the basis of
-#' dissimilarity with partitioning around medoids.
+#' dissimilarity with partitioning around medoids, using the Clustering Large
+#' Applications (CLARA) algorithm.
 #'
 #' @param dissimilarity the output object from [dissimilarity()] or
 #' [similarity_to_dissimilarity()], or a `dist` object. If a `data.frame` is
@@ -14,26 +15,25 @@
 #' @param n_clust an `integer` or a `vector` of `integers` specifying the
 #' requested number(s) of clusters.
 #' 
-#' @param variant a `character` string specifying the variant of pam to use,
-#' by default "faster". Available options are original, o_1, o_2, f_3, f_4,
-#' f_5 or fasterSee [cluster::pam()][cluster::pam] for more details.
+#' @param maxiter an `integer` defining the maximum number of iterations.
 #' 
-#' @param nstart an `integer` specifying the number of random “starts” for the
-#' pam algorithm. By default, 1 (for the `"faster"` variant).
+#' @param initializer character string, either 'BUILD' (used in classic PAM
+#' algorithm) or 'LAB' (linear approximative BUILD).
 #' 
-#' @param cluster_only a `boolean` specifying if only the clustering should be
-#' returned from the [cluster::pam()][cluster::pam] function (more efficient).
+#' @param fasttol Positive numeric defining the tolerance for fast swapping
+#' behavior, set to 1 by default.
 #' 
-#' @param ... you can add here further arguments to be passed to `pam()`
-#' (see [cluster::pam()][cluster::pam])
-#'
+#' @param numsamples Positive integer defining the number of samples to draw.
+#' 
+#' @param sampling Positive numeric defining the sampling rate.
+#' 
+#' @param independent Logical, FALSE by default meaning that the previous
+#' medoids are not kept in the next sample.
+#' 
+#' @param seed an `integer` to define a generator of random numbers.
+#' 
 #' @details
-#' This method partitions data into the chosen number of cluster on the basis
-#' of the input dissimilarity matrix. It is more robust than k-means because it
-#' minimizes the sum of dissimilarity between cluster centres and points
-#' assigned to the cluster - whereas the k-means approach minimizes the sum of
-#' squared euclidean distances (thus k-means cannot be applied directly on the
-#' input dissimilarity matrix if the distances are not euclidean).
+#' Based on fastkmedoids R package.
 #'
 #' @return
 #' A `list` of class `bioregion.clusters` with five slots:
@@ -46,14 +46,14 @@
 #' \item{**clusters**: `data.frame` containing the clustering results}}
 #' 
 #' @references
-#' \insertRef{Kaufman2009}{bioregion}
+#' \insertRef{Schubert2019}{bioregion}
 #' 
 #' @author
-#' Boris Leroy (\email{leroy.boris@gmail.com}),
-#' Pierre Denelle (\email{pierre.denelle@gmail.com}) and
+#' Pierre Denelle (\email{pierre.denelle@gmail.com}),
+#' Boris Leroy (\email{leroy.boris@gmail.com}), and
 #' Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) 
 #' 
-#' @seealso [nhclu_kmeans] 
+#' @seealso [nhclu_pam] 
 #' 
 #' @examples
 #' comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
@@ -61,30 +61,29 @@
 #' rownames(comat) <- paste0("Site",1:20)
 #' colnames(comat) <- paste0("Species",1:25)
 #'
-#' comnet <- mat_to_net(comat)
 #' dissim <- dissimilarity(comat, metric = "all")
+#'
+#' clust1 <- nhclu_clara(dissim, index = "Simpson", n_clust = 5)
 #' 
-#' clust1 <- nhclu_pam(dissim, n_clust = 2:10, index = "Simpson")
-#' clust2 <- nhclu_pam(dissim, n_clust = 2:15, index = "Simpson")
-#' partition_metrics(clust2, dissimilarity = dissim,
+#' partition_metrics(clust1, dissimilarity = dissim,
 #' eval_metric = "pc_distance")
-#' partition_metrics(clust2, net = comnet, species_col = "Node2",
-#'                    site_col = "Node1", eval_metric = "avg_endemism")
+#' 
 #'    
 #' @importFrom stats as.dist
-#' @importFrom cluster pam    
+#' @importFrom fastkmedoids fastclara    
 #'                    
 #' @export
 
-nhclu_pam <- function(
-    dissimilarity,
-    index = names(dissimilarity)[3],
-    n_clust = NULL,
-    nstart = if(variant == "faster") 1 else NA,
-    variant = "faster", # c("original","o_1","o_2","f_3","f_4","f_5","faster")
-    cluster_only = FALSE, # To reduce computation time & memory, can be
-    # provided to cluster functions
-    ...){ # Further arguments to be passed to cluster::pam
+nhclu_clara <- function(dissimilarity,
+                        index = names(dissimilarity)[3],
+                        n_clust = NULL,
+                        maxiter = 0L,
+                        initializer = "LAB",
+                        fasttol = 1,
+                        numsamples = 5L,
+                        sampling = 0.25,
+                        independent = FALSE,
+                        seed = 123456789L){
   
   # 1. Controls ---------------------------------------------------------------
   if(inherits(dissimilarity, "bioregion.pairwise.metric")){
@@ -137,25 +136,31 @@ nhclu_pam <- function(
     dist.obj <- dissimilarity
   }
   
-  if(!is.character(variant) || length(variant) != 1 ||
-     !(all(variant %in% c("original", "o_1", "o_2", "f_3", "f_4", "f_5",
-                          "faster")))){
-    stop("variant is a character string indicating. Available options are
-         original, o_1, o_2, f_3, f_4, f_5 or faster.")
+  controls(args = maxiter, data = NULL, type = "positive_integer")
+  if(!(initializer %in% c("BUILD", "LAB"))){
+    stop("initializer must be either 'BUILD' (used in classic PAM) or 'LAB'
+         (linear approximative BUILD).")
   }
   
-  controls(args = cluster_only, data = NULL, type = "boolean")
-  controls(args = nstart, data = NULL, type = "positive_integer")
-
+  controls(args = fasttol, data = NULL, type = "positive_numeric")
+  controls(args = numsamples, data = NULL, type = "positive_integer")
+  controls(args = sampling, data = NULL, type = "positive_numeric")
+  controls(args = independent, data = NULL, type = "boolean")
+  controls(args = seed, data = NULL, type = "positive_integer")
+  
   # 2. Function ---------------------------------------------------------------
-  outputs <- list(name = "pam")
+  # Output format
+  outputs <- list(name = "clara")
   
   outputs$args <- list(index = index,
                        n_clust = n_clust,
-                       nstart = nstart,
-                       variant = variant,
-                       cluster_only = cluster_only,
-                       ...)
+                       maxiter = maxiter,
+                       initializer = initializer,
+                       fasttol = fasttol,
+                       numsamples = numsamples,
+                       sampling = sampling,
+                       independent = independent,
+                       seed = seed)
   
   outputs$inputs <- list(bipartite = FALSE,
                          weight = TRUE,
@@ -173,24 +178,27 @@ nhclu_pam <- function(
   
   outputs$clusters$name <- labels(dist.obj)
   
-  outputs$algorithm$pam <- lapply(n_clust,
-                                  function(x)
-                                    cluster::pam(dist.obj,
-                                                 k = x,
-                                                 diss = TRUE,
-                                                 keep.diss = FALSE,
-                                                 keep.data = FALSE,
-                                                 nstart = nstart,
-                                                 variant = variant,
-                                                 cluster.only = cluster_only,
-                                                 ...))
+  # CLARA algorithm
+  outputs$algorithm$clara <-
+    lapply(n_clust,
+           function(x)
+             fastkmedoids::fastclara(rdist = dist.obj,
+                                     n = nrow(dist.obj),
+                                     k = x,
+                                     maxiter = maxiter,
+                                     initializer = initializer,
+                                     fasttol = fasttol,
+                                     numsamples = numsamples,
+                                     sampling = sampling,
+                                     independent = independent,
+                                     seed = seed))
   
-  names(outputs$algorithm$pam) <- paste0("K_", n_clust)
+  names(outputs$algorithm$clara) <- paste0("K_", n_clust)
   
   outputs$clusters <- data.frame(
     outputs$clusters,
-    data.frame(lapply(names(outputs$algorithm$pam),
-                      function(x) outputs$algorithm$pam[[x]]$clustering)))
+    data.frame(lapply(names(outputs$algorithm$clara),
+                      function(x) outputs$algorithm$clara[[x]]@assignment)))
   
   outputs$clusters <- knbclu(outputs$clusters, reorder = FALSE)