+++ R/nhclu_clarans.R
@@ -1,7 +1,8 @@
-#' Non hierarchical clustering: partitioning around medoids
+#' Non hierarchical clustering: CLARANS
 #'
 #' This function performs non hierarchical clustering on the basis of
-#' dissimilarity with partitioning around medoids.
+#' dissimilarity with partitioning around medoids, using the Clustering Large
+#' Applications based on RANdomized Search (CLARANS) algorithm.
 #'
 #' @param dissimilarity the output object from [dissimilarity()] or
 #' [similarity_to_dissimilarity()], or a `dist` object. If a `data.frame` is
@@ -14,26 +15,14 @@
 #' @param n_clust an `integer` or a `vector` of `integers` specifying the
 #' requested number(s) of clusters.
 #' 
-#' @param variant a `character` string specifying the variant of pam to use,
-#' by default "faster". Available options are original, o_1, o_2, f_3, f_4,
-#' f_5 or fasterSee [cluster::pam()][cluster::pam] for more details.
+#' @param numlocal an `integer` defining the number of samples to draw.
 #' 
-#' @param nstart an `integer` specifying the number of random “starts” for the
-#' pam algorithm. By default, 1 (for the `"faster"` variant).
+#' @param maxneighbor A positive numeric defining the sampling rate.
 #' 
-#' @param cluster_only a `boolean` specifying if only the clustering should be
-#' returned from the [cluster::pam()][cluster::pam] function (more efficient).
+#' @param seed an `integer` to define a generator of random numbers.
 #' 
-#' @param ... you can add here further arguments to be passed to `pam()`
-#' (see [cluster::pam()][cluster::pam])
-#'
 #' @details
-#' This method partitions data into the chosen number of cluster on the basis
-#' of the input dissimilarity matrix. It is more robust than k-means because it
-#' minimizes the sum of dissimilarity between cluster centres and points
-#' assigned to the cluster - whereas the k-means approach minimizes the sum of
-#' squared euclidean distances (thus k-means cannot be applied directly on the
-#' input dissimilarity matrix if the distances are not euclidean).
+#' Based on fastkmedoids R package.
 #'
 #' @return
 #' A `list` of class `bioregion.clusters` with five slots:
@@ -46,14 +35,14 @@
 #' \item{**clusters**: `data.frame` containing the clustering results}}
 #' 
 #' @references
-#' \insertRef{Kaufman2009}{bioregion}
+#' \insertRef{Schubert2019}{bioregion}
 #' 
 #' @author
-#' Boris Leroy (\email{leroy.boris@gmail.com}),
-#' Pierre Denelle (\email{pierre.denelle@gmail.com}) and
+#' Pierre Denelle (\email{pierre.denelle@gmail.com}),
+#' Boris Leroy (\email{leroy.boris@gmail.com}), and
 #' Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) 
 #' 
-#' @seealso [nhclu_kmeans] 
+#' @seealso [nhclu_pam] 
 #' 
 #' @examples
 #' comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
@@ -61,30 +50,25 @@
 #' rownames(comat) <- paste0("Site",1:20)
 #' colnames(comat) <- paste0("Species",1:25)
 #'
-#' comnet <- mat_to_net(comat)
 #' dissim <- dissimilarity(comat, metric = "all")
+#'
+#' clust1 <- nhclu_clarans(dissim, index = "Simpson", n_clust = 5)
 #' 
-#' clust1 <- nhclu_pam(dissim, n_clust = 2:10, index = "Simpson")
-#' clust2 <- nhclu_pam(dissim, n_clust = 2:15, index = "Simpson")
-#' partition_metrics(clust2, dissimilarity = dissim,
+#' partition_metrics(clust1, dissimilarity = dissim,
 #' eval_metric = "pc_distance")
-#' partition_metrics(clust2, net = comnet, species_col = "Node2",
-#'                    site_col = "Node1", eval_metric = "avg_endemism")
+#' 
 #'    
 #' @importFrom stats as.dist
-#' @importFrom cluster pam    
+#' @importFrom fastkmedoids fastclarans    
 #'                    
 #' @export
 
-nhclu_pam <- function(
-    dissimilarity,
-    index = names(dissimilarity)[3],
-    n_clust = NULL,
-    nstart = if(variant == "faster") 1 else NA,
-    variant = "faster", # c("original","o_1","o_2","f_3","f_4","f_5","faster")
-    cluster_only = FALSE, # To reduce computation time & memory, can be
-    # provided to cluster functions
-    ...){ # Further arguments to be passed to cluster::pam
+nhclu_clarans <- function(dissimilarity,
+                          index = names(dissimilarity)[3],
+                          n_clust = NULL,
+                          numlocal = 2L,
+                          maxneighbor = 0.025,
+                          seed = 123456789L){
   
   # 1. Controls ---------------------------------------------------------------
   if(inherits(dissimilarity, "bioregion.pairwise.metric")){
@@ -137,25 +121,19 @@ nhclu_pam <- function(
     dist.obj <- dissimilarity
   }
   
-  if(!is.character(variant) || length(variant) != 1 ||
-     !(all(variant %in% c("original", "o_1", "o_2", "f_3", "f_4", "f_5",
-                          "faster")))){
-    stop("variant is a character string indicating. Available options are
-         original, o_1, o_2, f_3, f_4, f_5 or faster.")
-  }
+  controls(args = numlocal, data = NULL, type = "positive_integer")
+  controls(args = maxneighbor, data = NULL, type = "positive_numeric")
+  controls(args = seed, data = NULL, type = "positive_integer")
   
-  controls(args = cluster_only, data = NULL, type = "boolean")
-  controls(args = nstart, data = NULL, type = "positive_integer")
-
   # 2. Function ---------------------------------------------------------------
-  outputs <- list(name = "pam")
+  # Output format
+  outputs <- list(name = "clarans")
   
   outputs$args <- list(index = index,
                        n_clust = n_clust,
-                       nstart = nstart,
-                       variant = variant,
-                       cluster_only = cluster_only,
-                       ...)
+                       numlocal = numlocal,
+                       maxneighbor = maxneighbor,
+                       seed = seed)
   
   outputs$inputs <- list(bipartite = FALSE,
                          weight = TRUE,
@@ -173,24 +151,23 @@ nhclu_pam <- function(
   
   outputs$clusters$name <- labels(dist.obj)
   
-  outputs$algorithm$pam <- lapply(n_clust,
-                                  function(x)
-                                    cluster::pam(dist.obj,
-                                                 k = x,
-                                                 diss = TRUE,
-                                                 keep.diss = FALSE,
-                                                 keep.data = FALSE,
-                                                 nstart = nstart,
-                                                 variant = variant,
-                                                 cluster.only = cluster_only,
-                                                 ...))
+  # CLARANS algorithm
+  outputs$algorithm$clarans <-
+    lapply(n_clust,
+           function(x)
+             fastkmedoids::fastclarans(rdist = dist.obj,
+                                       n = nrow(dist.obj),
+                                       k = x,
+                                       numlocal = numlocal,
+                                       maxneighbor = maxneighbor,
+                                       seed = seed))
   
-  names(outputs$algorithm$pam) <- paste0("K_", n_clust)
+  names(outputs$algorithm$clarans) <- paste0("K_", n_clust)
   
   outputs$clusters <- data.frame(
     outputs$clusters,
-    data.frame(lapply(names(outputs$algorithm$pam),
-                      function(x) outputs$algorithm$pam[[x]]$clustering)))
+    data.frame(lapply(names(outputs$algorithm$clarans),
+                      function(x) outputs$algorithm$clarans[[x]]@assignment)))
   
   outputs$clusters <- knbclu(outputs$clusters, reorder = FALSE)