+++ inst/doc/a1_install_binary_files.R
@@ -1,3 +1,0 @@
-## ----install_binaries, eval=FALSE---------------------------------------------
-#  install_binaries(binpath = "tempdir" , infomap_version = c("2.1.0", "2.6.0"))
-